move fn read_to_end to request_response crate
