+++ core/src/upgrade.rs
@@ -77,9 +77,7 @@ pub use self::{
     pending::PendingUpgrade,
     ready::ReadyUpgrade,
     select::SelectUpgrade,
-    transfer::{
-        read_length_prefixed, read_to_end, read_varint, write_length_prefixed, write_varint,
-    },
+    transfer::{read_length_prefixed, read_varint, write_length_prefixed, write_varint},
 };
 pub use crate::Negotiated;
 pub use multistream_select::{NegotiatedComplete, NegotiationError, ProtocolError, Version};
